should check course permission

+++ src/modules/course/helper/course.helper.ts
@@ -9,6 +9,7 @@ import { HandleServiceService } from "../../../modules/plan/services/handle_serv
 import { PlanService } from "../../../modules/plan/services/plan.service";
 import { User } from "../../../modules/user/schemas/user.schema";
 import { UserService } from "../../../modules/user/services/user.service";
+import { UserOrganizationService } from "../../../modules/user/services/user_organization.service";
 import { CreateCourseDto } from "../dto/create-course.dto";
 import { CreateCourseCalendarDto } from "../dto/create-course_calendar.dto";
 import {
@@ -73,7 +74,8 @@ export class CourseHelper {
     private planService: PlanService,
     private readonly eventHookNotificationService: EventHookNotificationService,
     private readonly hookWorker: EventHookWorkerService,
-    private readonly userService: UserService
+    private readonly userService: UserService,
+    private readonly userOrganization: UserOrganizationService
   ) {
     setTimeout(async () => {
       //await this.handleProcessModuleCount()
@@ -107,6 +109,11 @@ export class CourseHelper {
    */
   async createNewCourse(createCourseData: CreateCourseDto, res: Response, req: ExpressRequestDto) {
     try {
+      if (createCourseData.organization_id) {
+        const organization = await this.userOrganization.findOne({ _id: createCourseData.organization_id });
+        if (!organization) throw new Error("Not found Organization");
+      }
+
       createCourseData = { ...createCourseData, ...{ user_id: req.user_object._id.toString() } };
       let dataCreate: any = await this.courseService.create(createCourseData);
       let dataReturn: any = await this.courseService.findById(dataCreate?._id?.toString());
@@ -256,6 +263,9 @@ export class CourseHelper {
    */
   async updateCourse(dataUpdate: UpdateCourseDto, res: Response, req: ExpressRequestDto) {
     try {
+      const isValidUser = await this.checkUserCoursePermission(dataUpdate._id, req, res);
+      if (!isValidUser) throw new Error("You can't do this action since you're not a part of organization");
+
       let dataCreate: any = await this.courseService.update(dataUpdate);
       if (dataCreate?.price && !dataCreate?.service_id) {
         dataCreate = await this.handleUpdateServiceCourse(dataCreate);
@@ -810,8 +820,10 @@ export class CourseHelper {
    */
   async handleDeleteCourse(id: string, res: Response, req: ExpressRequestDto) {
     try {
-      let dataReturn = await this.courseService.remove(id);
+      const isValidUser = await this.checkUserCoursePermission(id, req, res);
+      if (!isValidUser) throw new Error("You can't do this action since you're not a part of organization");
 
+      let dataReturn = await this.courseService.remove(id);
       if (!dataReturn) throw new NotFoundException("Not found course");
 
       return res
@@ -917,6 +929,9 @@ export class CourseHelper {
    */
   async handleAddUserToCourse(dataFollow: CreateCourseUserDto, req: ExpressRequestDto, res: Response) {
     try {
+      const isValidUser = await this.checkUserCoursePermission(dataFollow.course_id, req, res);
+      if (!isValidUser) throw new Error("You can't do this action since you're not a part of organization");
+
       let userObject = req?.user_object;
       let userId = userObject._id;
 
@@ -1286,11 +1301,8 @@ export class CourseHelper {
       const userObject = req.user_object;
       if (!userObject) throw new Error("User is invalid");
 
-      const course = await this.courseService.findOne({
-        user_id: userObject._id.toString(),
-        _id: dataFollow.course_id,
-      });
-      if (!course) throw new Error("User don't have permission to create new class in this course");
+      const isValidUser = await this.checkUserCoursePermission(dataFollow.course_id, req, res);
+      if (!isValidUser) throw new Error("You can't do this action since you're not a part of this course");
 
       // check other class time
       const courseClasses = await this.courseClassService.getAllAssignedTimeInCourse(dataFollow.course_id);
@@ -1352,6 +1364,9 @@ export class CourseHelper {
         _id: dataFollow._id,
       });
       if (!oldClass) throw new Error("Not found your class");
+
+      const isValidUser = await this.checkUserCoursePermission(oldClass.course_id.toString(), req, res);
+      if (!isValidUser) throw new Error("You can't do this action since you're not a part of organization");
 
       const courseInfo = await this.courseService.findOne({ _id: oldClass.course_id.toString() });
       const isAfter = moment().isAfter(moment(courseInfo.start_time.toString()));
@@ -1424,6 +1439,14 @@ export class CourseHelper {
 
   async deleteClass(id: string, req: ExpressRequestDto, res: Response) {
     try {
+      const oldClass = await this.courseClassService.findOne({
+        _id: id,
+      });
+      if (!oldClass) throw new Error("Not found your class");
+
+      const isValidUser = await this.checkUserCoursePermission(oldClass.course_id.toString(), req, res);
+      if (!isValidUser) throw new Error("You can't do this action since you're not a part of organization");
+
       const courseClass = await this.courseClassService.remove({ _id: id });
 
       return res
@@ -1441,6 +1464,9 @@ export class CourseHelper {
         _id: dataFollow.class_id,
       });
       if (!courseClass) throw new Error("Not found your class");
+
+      const isValidUser = await this.checkUserCoursePermission(courseClass.course_id.toString(), req, res);
+      if (!isValidUser) throw new Error("You can't do this action since you're not a part of organization");
 
       const isUserBoughtCourse = await this.courseUserService.findOne({
         user_id: dataFollow.user_id,
@@ -1471,6 +1497,9 @@ export class CourseHelper {
         _id: dataFollow.class_id,
       });
       if (!courseClass) throw new Error("Not found your class");
+
+      const isValidUser = await this.checkUserCoursePermission(courseClass.course_id.toString(), req, res);
+      if (!isValidUser) throw new Error("You can't do this action since you're not a part of organization");
 
       const newMembers = courseClass.members.filter((member) => {
         return member.toString() !== dataFollow.user_id;
@@ -1886,5 +1915,20 @@ export class CourseHelper {
       return accumulator + currentValue.rating;
     }, 0);
     return totalRating / reviews.length;
+  }
+
+  async checkUserCoursePermission(courseId: string, req: ExpressRequestDto, res: Response): Promise<boolean> {
+    const course = await this.courseService.findOne({ _id: courseId });
+    if (course?.organization_id) {
+      if (req.user_object?.organization_id.toString() === course?.organization_id.toString()) {
+        return true;
+      } else return false;
+    } else if (course?.user_id) {
+      if (req.user_object?._id.toString() === course?.user_id?._id.toString()) {
+        return true;
+      } else return false;
+    }
+
+    return true;
   }
 }
